New image for Agda 2.6.0, stdlib 1.0
This image uses a different method for building Agda. Previously, we
would get Agda from Stackage. However, it always takes a little while
for Agda to land in a Stackage snapshot, so now we get the sources from
Github and build them directly. This change unfortunately increases the
image size by a few megabytes.